Refactor: total 모듈을 stat 모듈로 합침
