+++ src/auth/auth.module.ts
@@ -5,13 +5,13 @@ import { JwtConfig } from '../Config/jwt.config';
 import { JwtModule } from '@nestjs/jwt';
 import { UserModule } from '../user/user.module';
 import { JwtStrategy } from './strategy/jwt.strategy';
-import { TotalModule } from '../total/total.module';
+import { StatModule } from '../stat/stat.module';
 
 @Module({
     imports: [
         JwtModule.registerAsync({ useClass: JwtConfig }),
         UserModule,
-        TotalModule,
+        StatModule,
     ],
     controllers: [AuthController],
     providers: [AuthService, JwtStrategy],
