+++ src/stat/stat.module.ts
@@ -6,6 +6,8 @@ import { AlgorithmService } from './service/algorithm.service';
 import { AlgorithmRepository } from './repository/algorithm.repository';
 import { GradeService } from './service/grade.service';
 import { GradeRepository } from './repository/grade.repository';
+import { TotalService } from './service/total.service';
+import { TotalRepository } from './repository/total.repository';
 
 @Module({
     controllers: [StatController],
@@ -16,6 +18,9 @@ import { GradeRepository } from './repository/grade.repository';
         AlgorithmRepository,
         GradeService,
         GradeRepository,
+        TotalService,
+        TotalRepository,
     ],
+    exports: [TotalService],
 })
 export class StatModule {}
